+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Improvements
 	* The `--dry-run` option now prevents from persisting any configuration, including pending shows, not just the last execution date.
+	* Performance improvements when running with the `-t` flag.
 
 ## 2.6.5 (2021-06-10)