+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # download_tv CHANGELOG
 
-## Unreleased (????-??-??)
+## 2.6.6 (2022-01-21)
 
 * Improvements
 	* The `--dry-run` option now prevents from persisting any configuration, including pending shows, not just the last execution date.